Changed to read RSS feeds in background thread.
The program processes the file $XDG_CONFIG_HOME/gtk-feed/feeds (usually
~/.config/gtk-feed/feeds) at startup, which should contain a list of RSS
feed URIs to read.  For each feed URI, a background thread is spawned and
the thread constructs a submenu corresponding to the contents of the feed.
Thread synchronization may not be properly handled yet; more research and
work needs to be done on this and the program may be unstable.